+++ src/components/voting/ModalVideoContent.js
