+++ src/components/voting/member-overview/MemberInfo.js
