+++ src/components/voting/member-overview/RecentActivities.js
